+++ src/templates/template.h
@@ -333,7 +333,7 @@ public:
 	 * The scale is the combined view & template scale. It can be used to give
 	 * a minimum size to elements.
 	 */
-    virtual void drawTemplate(QPainter* painter, const QRectF& clip_rect, double scale, bool on_screen, float opacity) const = 0;
+    virtual void drawTemplate(QPainter* painter, const QRectF& clip_rect, double scale, bool on_screen, qreal opacity) const = 0;
 	
 	
 	/**
