+++ src/templates/template_map.h
@@ -25,6 +25,7 @@
 #include <memory>
 #include <vector>
 
+#include <QtGlobal>
 #include <QObject>
 #include <QString>
 
@@ -71,7 +72,7 @@ public:
 	void unloadTemplateFileImpl() override;
 	
 	
-	void drawTemplate(QPainter* painter, const QRectF& clip_rect, double scale, bool on_screen, float opacity) const override;
+	void drawTemplate(QPainter* painter, const QRectF& clip_rect, double scale, bool on_screen, qreal opacity) const override;
 	
 	QRectF getTemplateExtent() const override;
